fix recreation of indexes

+++ store/postgres/src/deployment_store.rs
@@ -2,6 +2,7 @@ use detail::DeploymentDetail;
 use diesel::connection::SimpleConnection;
 use diesel::pg::PgConnection;
 use diesel::r2d2::{ConnectionManager, PooledConnection};
+use diesel::sql_types::{Bool, Text};
 use diesel::{prelude::*, sql_query};
 use graph::anyhow::Context;
 use graph::blockchain::block_stream::FirehoseCursor;
@@ -1495,13 +1496,13 @@ impl DeploymentStore {
         site: Arc<Site>,
         graft_src: Option<(Arc<Layout>, BlockPtr, SubgraphDeploymentEntity, IndexList)>,
     ) -> Result<(), StoreError> {
-        // #[derive(QueryableByName, Debug)]
-        // struct IndexInfo {
-        //     #[diesel(sql_type = Bool)]
-        //     isvalid: bool,
-        //     #[diesel(sql_type = Bool)]
-        //     isready: bool,
-        // }
+        #[derive(QueryableByName, Debug)]
+        struct IndexInfo {
+            #[diesel(sql_type = Bool)]
+            isvalid: bool,
+            #[diesel(sql_type = Bool)]
+            isready: bool,
+        }
 
         let dst = self.find_layout(site.cheap_clone())?;
 
@@ -1609,47 +1610,55 @@ impl DeploymentStore {
 
         // check if all indexes are valid and recreate them if they aren't
         let mut conn = self.get_conn()?;
-        // TODO: fix it
-        /*
+
         if ENV_VARS.postpone_attribute_index_creation {
             let namespace = site.namespace.as_str();
+            let index_list = self.load_indexes(site.clone())?;
             for table in dst.tables.values() {
-                for (index_name, create_query) in table.create_postponed_indexes() {
-                    let table_name = table.name.clone();
-                    let query = r#"
-                        SELECT  x.indisvalid           AS isvalid,
-                                x.indisready           AS isready
-                        FROM pg_index x
-                                JOIN pg_class c ON c.oid = x.indrelid
-                                JOIN pg_class i ON i.oid = x.indexrelid
-                                LEFT JOIN pg_namespace n ON n.oid = c.relnamespace
-                        WHERE (c.relkind = ANY (ARRAY ['r'::"char", 'm'::"char", 'p'::"char"]))
-                        AND (i.relkind = ANY (ARRAY ['i'::"char", 'I'::"char"]))
-                        AND (n.nspname = $1)
-                        AND (c.relname = $2)
-                        AND (i.relname = $3);"#;
-                    let ii_vec = sql_query(query)
-                        .bind::<Text, _>(namespace)
-                        .bind::<Text, _>(table_name)
-                        .bind::<Text, _>(index_name.clone())
-                        .get_results::<IndexInfo>(&mut conn)?
-                        .into_iter()
-                        .map(|ii| ii.into())
-                        .collect::<Vec<IndexInfo>>();
-                    assert!(ii_vec.len() <= 1);
-                    // Check if the index is valid. If either isvalid or isready flag of pg_index table
-                    // isn't true, drop it and recreate it.
-                    if ii_vec.len() == 0 || !ii_vec[0].isvalid || !ii_vec[0].isready {
-                        if ii_vec.len() > 0 {
-                            let drop_query =
-                                sql_query(format!("DROP INDEX {}.{};", namespace, index_name));
-                            conn.transaction(|conn| drop_query.execute(conn))?;
+                for (ind_name, create_query) in index_list.indexes_for_table(
+                    &namespace.to_string(),
+                    &table.name.to_string(),
+                    table,
+                    true,
+                    true,
+                ) {
+                    if let Some(index_name) = ind_name {
+                        let table_name = table.name.clone();
+                        let query = r#"
+                            SELECT  x.indisvalid           AS isvalid,
+                                    x.indisready           AS isready
+                            FROM pg_index x
+                                    JOIN pg_class c ON c.oid = x.indrelid
+                                    JOIN pg_class i ON i.oid = x.indexrelid
+                                    LEFT JOIN pg_namespace n ON n.oid = c.relnamespace
+                            WHERE (c.relkind = ANY (ARRAY ['r'::"char", 'm'::"char", 'p'::"char"]))
+                            AND (i.relkind = ANY (ARRAY ['i'::"char", 'I'::"char"]))
+                            AND (n.nspname = $1)
+                            AND (c.relname = $2)
+                            AND (i.relname = $3);"#;
+                        let ii_vec = sql_query(query)
+                            .bind::<Text, _>(namespace)
+                            .bind::<Text, _>(table_name)
+                            .bind::<Text, _>(index_name.clone())
+                            .get_results::<IndexInfo>(&mut conn)?
+                            .into_iter()
+                            .map(|ii| ii.into())
+                            .collect::<Vec<IndexInfo>>();
+                        assert!(ii_vec.len() <= 1);
+                        // Check if the index is valid. If either isvalid or isready flag of pg_index table
+                        // isn't true, drop it and recreate it.
+                        if ii_vec.len() == 0 || !ii_vec[0].isvalid || !ii_vec[0].isready {
+                            if ii_vec.len() > 0 {
+                                let drop_query =
+                                    sql_query(format!("DROP INDEX {}.{};", namespace, index_name));
+                                conn.transaction(|conn| drop_query.execute(conn))?;
+                            }
+                            sql_query(create_query).execute(&mut conn)?;
                         }
-                        sql_query(create_query).execute(&mut conn)?;
                     }
                 }
             }
-        } */
+        }
 
         // Make sure the block pointer is set. This is important for newly
         // deployed subgraphs so that we respect the 'startBlock' setting
@@ -2091,7 +2100,7 @@ impl IndexList {
         dest_table: &Table,
         postponed: bool,
         concurent_if_not_exist: bool,
-    ) -> Vec<String> {
+    ) -> Vec<(Option<String>, String)> {
         let mut arr = vec![];
         if let Some(vec) = self.indexes.get(table_name) {
             for ci in vec {
@@ -2102,7 +2111,7 @@ impl IndexList {
                                 .with_nsp(namespace.clone())
                                 .to_sql(concurent_if_not_exist, concurent_if_not_exist)
                             {
-                                arr.push(sql)
+                                arr.push((ci.name(), sql))
                             }
                         }
                     }
